requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
SQLAlchemy-Utils==0.41.2
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
